python_scripts/requirements.txt
numpy
Pillow
python-dotenv
tifffile
webdavclient3
requests
argparse
